feat: implement JournalService interface and its implementation for journal management
